+++ R/data.R
@@ -9,7 +9,7 @@
 #' @format A list of character vectors indicating the standard
 #'   orders of row names in catalogs.
 #'
-#' @note In ID (insertion and deletion) catalogs, deletion repeat sizes
+#' @note In ID (small insertion and deletion) catalogs, deletion repeat sizes
 #'   range from 0 to 5+, but for plotting and end-user documentation
 #'   deletion repeat sizes range from 1 to 6+.
 #'