+++ tests/testthat/test_PlotCatalog.IndelCatalog.R
@@ -10,12 +10,12 @@ test_that("PlotCatalog.IndelCatalog function", {
                        region = "genome", catalog.type = "counts")
   cat <- catalog[, 1, drop = FALSE]
   out <- PlotCatalog(cat)
-  expect_equal(out, TRUE)
+  expect_equal(out$plot.success, TRUE)
 
   cat.counts.signature <- apply(cat, MARGIN = 2, function(x) x / sum(x))
   cat.counts.signature <-
     as.catalog(cat.counts.signature, ref.genome = "GRCh37",
                region = "genome", catalog.type = "counts.signature")
   out <- PlotCatalog(cat.counts.signature)
-  expect_equal(out, TRUE)
+  expect_equal(out$plot.success, TRUE)
 })