+++ tests/testthat/test_PlotCatalog.SBS192Catalog.R
@@ -12,8 +12,8 @@ test_that("PlotCatalog.SBS192Catalog function", {
   cat.counts <- catalog.counts[, 1, drop = FALSE]
   out <- PlotCatalog(cat.counts)
   out1 <- PlotCatalog(cat.counts, plot.SBS12 = TRUE)
-  expect_equal(out, TRUE)
-  expect_equal(out1, TRUE)
+  expect_equal(out$plot.success, TRUE)
+  expect_equal(out1$plot.success, TRUE)
   
   cat.density <-
     TransformCatalog(cat.counts, target.ref.genome = "GRCh37",
@@ -21,8 +21,8 @@ test_that("PlotCatalog.SBS192Catalog function", {
                      target.catalog.type = "density")
   out <- PlotCatalog(cat.density)
   out1 <- PlotCatalog(cat.density, plot.SBS12 = TRUE)
-  expect_equal(out, TRUE)
-  expect_equal(out1, TRUE)
+  expect_equal(out$plot.success, TRUE)
+  expect_equal(out1$plot.success, TRUE)
 
   cat.counts.signature <-
     TransformCatalog(cat.counts, target.ref.genome = "GRCh37",
@@ -30,8 +30,8 @@ test_that("PlotCatalog.SBS192Catalog function", {
                      target.catalog.type = "counts.signature")
   out <- PlotCatalog(cat.counts.signature)
   out1 <- PlotCatalog(cat.counts.signature, plot.SBS12 = TRUE)
-  expect_equal(out, TRUE)
-  expect_equal(out1, TRUE)
+  expect_equal(out$plot.success, TRUE)
+  expect_equal(out1$plot.success, TRUE)
 
   cat.density.signature <-
     TransformCatalog(cat.counts, target.ref.genome = "GRCh37",
@@ -39,6 +39,6 @@ test_that("PlotCatalog.SBS192Catalog function", {
                      target.catalog.type = "density.signature")
   out <- PlotCatalog(cat.density.signature)
   out1 <- PlotCatalog(cat.density.signature, plot.SBS12 = TRUE)
-  expect_equal(out, TRUE)
-  expect_equal(out1, TRUE)
+  expect_equal(out$plot.success, TRUE)
+  expect_equal(out1$plot.success, TRUE)
 })